Migrate PATCHES.json to new metadata schema
In an effort to automate patch synchronization between CrOS and
Android LLVM, we're moving the "comment" and "bugs_tests" fields
in PATCHES.json into a "metadata" dictionary entry instead.

Additionally, this sorts the key fields in PATCHES.json so that
keys won't be shuffled around as we apply new patches from different
programs.

Finally, this adds default values for every key to aid in
self-documenting the schema.

BUG=b:209493133
TEST=sudo emerge sys-devel/llvm
TEST=CQ

Change-Id: If9c4e4c9d9244833ed62c56fd1ac2fba0121e2f8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3319892
Tested-by: Jordan Abrahams <ajordanr@google.com>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Commit-Queue: Jordan Abrahams <ajordanr@google.com>