cros-workon: On error for -9999.ebuilds check if rebase needed.
If a developer's checkout is not consistent they can get odd errors that
manifest differently for each case. Since these are hard to troubleshoot
unless you know what to look for, test for this case and print an
actionable error message.

This also cleans up the following:
* some missing local variable statements.
* an unused variable.

BUG=None
TEST=manually test error case with and without message.

Change-Id: I1dea96a5c004b7a342349be989246843fefac629
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3331435
Tested-by: Allen Webb <allenwebb@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Allen Webb <allenwebb@google.com>