crosvm: Remove -r1 ebuild files
Because CROS_WORKON_MANUAL_UPREV was removed on the unstable ebuild file
but not on the stable one, the annealing process ignored the existing
stable ebuild file and created a new -r1 revision.

See: https://crrev.com/c/3336409

This currently breaks all changes to crosvm in the CQ, as the next uprev created will
be -r2, but the existing higher revision will be used in the build and
not contain the code to be tested.

BUG=b:209034086
TEST=None

Change-Id: If88da7dd3e7ab19f2bf1539b9f407949258d6922
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3343874
Commit-Queue: Dennis Kempin <denniskempin@google.com>
Tested-by: Dennis Kempin <denniskempin@google.com>
Auto-Submit: Dennis Kempin <denniskempin@google.com>
Reviewed-by: Junichi Uekawa <uekawa@chromium.org>
Commit-Queue: Junichi Uekawa <uekawa@chromium.org>