dhcpcd: Additional checks for ARP packets
This is yet another attempt to fix crashes reported in arp_read().  The
problem that it tries to address is that normally ARP packet checks are
performed by the BPF filter, but there is a short time interval between
opening of socket and installing of the filter where we might get some
packets.  So in addition to the BPF checks (which should be more
efficient in filtering out non-ARP packets) we also do some checks in
arp_packet() function that is called from arp_read().

BUG=b:204389871
TEST=emerge-${BOARD} dhcpcd, deploy, and check if wifi_matfunc and
     network_nightly are passing.

Change-Id: I351c1fe4640ce16746588638ae41d09c0bc860f1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3347795
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Tested-by: Andrzej Ostruszka <andrzejo@google.com>
Commit-Queue: Andrzej Ostruszka <andrzejo@google.com>