brask: add the libpayload config for turning off CONFIG_LP_PC_I8042
The brask doesn't need the i8042 driver. To avoid the autotest
error, we add the brask libpayload to turn off the CONFIG_LP_PC_I8042.

Note: Since we use the libpayload name to find the baseboard config
(ie. config.baseboard.LIBPAYLOAD), we need to add the link for the
config.baseboard.brask to avoid the build error.

BUG=b:207077409
TEST=`test_that power_Resume` passed

Change-Id: Iceeb8bddb41ce13c8ac25c89fb335b48d2bae0cd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3353305
Tested-by: Zhuohao Lee <zhuohao@chromium.org>
Reviewed-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Commit-Queue: Zhuohao Lee <zhuohao@chromium.org>