brya: Add agah coreboot configuration
Create a new coreboot configuration for the agah variant
of the brya0 reference board. The new configuration file is
a copy of the reference board, but the name of the baseboard is
replaced by the name of the variant where applicable.

(Auto-Generated by create_coreboot_config.sh version 2.5.0).

BUG=b:210970640
TEST=FW_NAME=agah emerge-brya coreboot chromeos-bootimage
Ensure that image-agah.*.bin are created

Change-Id: Ibc7f658431b161342d3acc3ac72b371d21d72fab
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3372246
Tested-by: Tony Huang <tony-huang@quanta.corp-partner.google.com>
Reviewed-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Commit-Queue: Tim Wawrzynczak <twawrzynczak@chromium.org>