brya: Add volmar coreboot configuration
Create a new coreboot configuration for the volmar variant
of the brya0 reference board. The new configuration file is
a copy of the reference board, but the name of the baseboard is
replaced by the name of the variant where applicable.

(Auto-Generated by create_coreboot_config.sh version 2.5.0).

BUG=b:213127419
TEST=FW_NAME=volmar emerge-brya coreboot chromeos-bootimage
Ensure that image-volmar.*.bin are created

Cq-Depend: chrome-internal:4431700, chromium:3379421
Change-Id: I3c232ab8c0683fff051d4ab9e9d4901636cedb3f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3367805
Reviewed-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Tested-by: David Wu <david_wu@quanta.corp-partner.google.com>
Commit-Queue: David Wu <david_wu@quanta.corp-partner.google.com>