dev-rust/tokio: add 1.15.0 as an empty crate
Tokio 1.x is an optional dependency for dev-rust/gpio-cdev. Add it as an
empty crate to satisfy cargo's parser, while avoiding packaging the
large tree of dependencies it pulls in.

Note that the real Tokio 0.1.x stack remains in the tree for the
existing packages that use it. We can replace this empty crate with the
real Tokio 1.x later if that becomes necessary.

BUG=b:215622663
TEST=emerge-brya '=dev-rust/tokio-1*'

Change-Id: Icaf599c033329186d6fe305402314618ca96fbe1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3398362
Tested-by: Dan Callaghan <dcallagh@chromium.org>
Reviewed-by: David Lattimore <dml@chromium.org>
Commit-Queue: David Lattimore <dml@chromium.org>
Tested-by: David Lattimore <dml@chromium.org>