chromeos-base/crosvm: disable tests using /dev/kvm
The /dev/kvm device node may not have the appropriate permissions to be
accessed by the ebuild tests, and it may not be available at all if the
system does not have virtualization extensions enabled.

These tests are already executed in the Kokoro test infrastructure that
is enforced for all commits into the crosvm repository, so running them
again on the Chrome OS builders is not adding any extra coverage.

BUG=b:191974904
TEST=FEATURES=test emerge-nami crosvm

Change-Id: Id0d3e017d12e77e2dbcfa052bdae1d567acee75f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3381819
Reviewed-by: Dennis Kempin <denniskempin@google.com>
Tested-by: Daniel Verkamp <dverkamp@chromium.org>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>