brya: Add banshee coreboot configuration
Create a new coreboot configuration for the banshee variant
of the brya0 reference board. The new configuration file is
a copy of the reference board, but the name of the baseboard is
replaced by the name of the variant where applicable.

(Auto-Generated by create_coreboot_config.sh version 2.5.0).

BUG=b:214871796
TEST=FW_NAME=banshee emerge-brya coreboot chromeos-bootimage
Ensure that image-banshee.*.bin are created

Cq-Depend: chromium:3408052, chrome-internal:4474116, chrome-internal:4484347
Change-Id: Ie03f8c65980ce21a96a383c90230c3235b6cf4df
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3399378
Reviewed-by: YH Lin <yueherngl@chromium.org>
Tested-by: Nick Vaccaro <nvaccaro@google.com>
Reviewed-by: Nick Vaccaro <nvaccaro@google.com>
Reviewed-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Commit-Queue: Tim Wawrzynczak <twawrzynczak@chromium.org>