chromeos-base: Add udev rule for RTL8153-DD USB NIC
In order to support WOL, Realtek RTL8153-DD IC needs to change it's
config to match r8152 instead of cdc_ncm drivers.  r8152 driver includes
the chip specific code to support WOL for RTL8153-DD.  However, the
CDC class drivers, such as cdc_ether and cdc_ncm, do not.

The purpose of the patch is not to force the device into "vendor mode",
although it results in the effect. The maintainers of Linux kernel asked
that the user could change the configuration in runtime, and the driver
shouldn't limit it. That is, the user is able to use any mode (vendor,
ECM, or NCM), and it could be changed in runtime.

If cdc_ncm claims the device before udev rule is applied, cdc_ncm would be
unbound when configuration is changed. Then, r8152 will claim the device.

The rule is numbered to be applied after 51-mist.rules but before the
75-net-description.rules. This shouldn't significantly affect when the
driver can be loaded and initialize the device.  See preload-network.conf
and udev-trigger-early.conf upstart scripts for earlier loading and binding
of the networking (ethernet and wifi) drivers.

BUG=b:212196481
TEST=emerge-dedede chromeos-base/chromeos-base

Change-Id: Ifc6cf0ec92e8ca2df9b80e5758eefbc0d77c9af6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3401020
Tested-by: Grant Grundler <grundler@chromium.org>
Reviewed-by: Grant Grundler <grundler@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Grant Grundler <grundler@chromium.org>