x11-drivers/opengles-headers: Improve package dependency
This package simply install OpenGLES headers and it does not need other
packages in DEPEND to build.
Presumably the dependencies are added for transitive DEPEND, so packages
that depend on this (the "consumers") can also access the DEPENDed
headers e.g. the SPIR-V headers.

However if those dependencies are not shown in RDEPEND, they will not be
installed when this package is emerged from a binpkg, and the consumers
will fail to find the headers in that case.

See http://go/ebuild-faq#dependency-types for the details, or
b:199737686 for some discussion.
Basically we almost always want to list DEPEND in RDEPEND.

BUG=chromium:1288949
TEST=emerge-${BOARD} opengles-headers

Change-Id: If0f62ab31160a2c6f412e1be7b2f4d14fdc3eda2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3403459
Reviewed-by: Joshua Emele <jemele@chromium.org>
Tested-by: Fei Shao <fshao@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Fei Shao <fshao@chromium.org>