brya: Add crota coreboot configuration
Create a new coreboot configuration for the crota variant
of the brya0 reference board. The new configuration file is
a copy of the reference board, but the name of the baseboard is
replaced by the name of the variant where applicable.

(Auto-Generated by create_coreboot_config.sh version 2.5.0).

BUG=b:215443524
TEST=FW_NAME=crota emerge-brya coreboot chromeos-bootimage
Ensure that image-crota.*.bin are created

Cq-Depend: chromium:3419702
Change-Id: I06163239ba17d4d784d379d4f7b1180261f0449d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3420104
Tested-by: Terry Chen <terry_chen@wistron.corp-partner.google.com>
Reviewed-by: YH Lin <yueherngl@chromium.org>
Commit-Queue: Terry Chen <terry_chen@wistron.corp-partner.google.com>