chromeos-base/hps-firmware: delete committed Cargo.lock before compiling
The exact versions and checksums of the crates in the Chromium OS local
registry will not necessarily match what is in Cargo.lock, so we need to
delete Cargo.lock before trying to invoke cargo.

Normally this is handled by cros-rust_src_configure, but we didn't need
it before now because we didn't have Cargo.lock files committed. Now we
do.

BUG=b:201372907
TEST=FEATURES=test emerge-brya hps-firmware

Cq-Depend: chromium:3419029
Change-Id: Icf934c58ba06d7d0f6d460ce0a4934fdb3856d40
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3419616
Tested-by: Dan Callaghan <dcallagh@chromium.org>
Reviewed-by: David Lattimore <dml@chromium.org>
Commit-Queue: Dan Callaghan <dcallagh@chromium.org>