dev-rust/image: pull in the real num-iter crate
The actual dependency is on num-iter >= 0.1.32, as written. But we have
num-iter 0.1.37 as an empty crate, and 0.1.42 as the real crate.

Depend on 0.1.42 to ensure Portage picks the real crate and not the
empty one, particularly on developer's machines where they may still
have num-iter 0.1.37 installed.

BUG=none
TEST=emerge-brya dev-rust/image

Change-Id: I4c91646a80e183dbaa8368964803fd62fd630e49
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3418626
Tested-by: Dan Callaghan <dcallagh@chromium.org>
Reviewed-by: David Lattimore <dml@chromium.org>
Commit-Queue: Dan Callaghan <dcallagh@chromium.org>