eclass/cros-workon: Change hint language to be more clear.
A check was added in https://crrev.com/c/3331435 to print an error
message if a cros-workon package fails to emerge and one of its source
trees is out of sync. This updates the original language so that it is
more clear that the actual error message appeared earlier.

BUG=None
TEST=CQ passes

Change-Id: Iadc6e83ad05553965b2b38fb9b8de548dea15062
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3439623
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Allen Webb <allenwebb@google.com>
Tested-by: Allen Webb <allenwebb@google.com>