eclass/cros-rust: Verify tests are built before getting list.
Previously if a custom src_compile was used there is a chance the
developer might not have built the tests before running them. This
results in any compilation errors for the tests being hidden.

To resolve this the command was moved to make sure the tests are built
beforehand.

BUG=b:205999326
TEST=FEATURES=test emerge-${BOARD} <cros-rust cros-workon ebuild>

Change-Id: I3390d789c8ad3af592b800beec256604a3ac0533
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3439074
Reviewed-by: Dennis Kempin <denniskempin@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Allen Webb <allenwebb@google.com>
Commit-Queue: Allen Webb <allenwebb@google.com>