chromeos-base/hps-firmware: build private RISC-V GCC
Build a private copy of RISC-V binutils, GCC, and newlib, using the
toolchain scripts originally developed by Sifive. In a subsequent
commit, we will use the resulting C++ compiler to build Tensorflow Lite
for Microcontrollers.

For now we use the 2021.04.23 snapshot of the toolchain because it's the
version I have been testing with for nearly a year, and it provides
GCC 10.2 which almost (but not quite) matches the GCC 10.1 toolchain we
use in CFU-Playground.

BUG=b:199699065
TEST=emerge-brya hps-firmware

Change-Id: I5c385d74267d38fa23ce173cbba845981e8da5ad
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3439789
Tested-by: Dan Callaghan <dcallagh@chromium.org>
Reviewed-by: Alan Green <avg@chromium.org>
Commit-Queue: Dan Callaghan <dcallagh@chromium.org>