chromeos-base/hps-firmware: don't install stage0 image
The only copy of stage0 that matters is the one embedded in hps-factory.
Installing the image into the Chrome OS image serves no purpose, so
remove it.

We continue to build stage0 (and then discard the image) so that it's
tested in CQ. Eventually, it will be built as part of hps-factory
instead.

BUG=b:206034105
TEST=emerge-brya hps-firmware

Change-Id: I5c9413851081f0b0e9884d8c77382a9d124d3735
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3446221
Reviewed-by: David Lattimore <dml@chromium.org>
Tested-by: Dan Callaghan <dcallagh@chromium.org>
Commit-Queue: Dan Callaghan <dcallagh@chromium.org>