brask: Add moli coreboot configuration
Create a new coreboot configuration for the moli variant
of the brask reference board. The new configuration file is
a copy of the reference board, but the name of the baseboard is
replaced by the name of the variant where applicable.

(Auto-Generated by create_coreboot_config.sh version 2.5.0).

BUG=b:214439135
TEST=FW_NAME=moli emerge-brask coreboot chromeos-bootimage
Ensure that image-moli.*.bin are created

Cq-Depend: chromium:3456073
Change-Id: Ibc39c8787a6433d32f976710f56618b8713c6c56
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3449604
Tested-by: Raihow Shi <raihow_shi@wistron.corp-partner.google.com>
Reviewed-by: Zhuohao Lee <zhuohao@chromium.org>
Commit-Queue: Raihow Shi <raihow_shi@wistron.corp-partner.google.com>