chromeos-ec: Disable Kconfig checker from ebuild
The Kconfig checker is designed to keep config options in-sync between
the Zephyr build and the legacy EC build.  It runs via the Makefile,
and will cause the build to error when Kconfig differences exist.
This is to get the EC developers to fix the option and amend their
changes before re-uploading.

However, it has also recently been responsible for multiple blow-ups,
and as such, we should limit the scope to where it runs to the CLs
that might actually affect it.  The CLs that land in platform/ec do
not use this ebuild, and therefore this is a spot we should disable
it.

BUG=b:181323955
TEST=emerge-herobrine chromeos-ec passes

Change-Id: I5e5db22fa2c7a405d8ea89047845b0af21b6a0ec
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3450757
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Auto-Submit: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>