dev-embedded/hps-sdk: new package
This package is intended to contain compilers and tools for HPS firmware
targetting RISC-V, which is not supported in the main Chrome OS
toolchain. For now this package just builds Rust, but it will eventually
build a RISC-V C++ compiler too.

This ebuild is largely based on the existing dev-lang/rust ebuild, but
cut down to the bare minimum and simplified, with some inspiration taken
from dev-embedded/ti50-sdk.

This new package's installed size on disk is 307MB.

BUG=b:201365430
TEST=emerge hps-sdk

Change-Id: Ic8b46b6168d43f06b730ee8a0f815a640b8ff645
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3449597
Tested-by: Dan Callaghan <dcallagh@chromium.org>
Reviewed-by: Alan Green <avg@chromium.org>
Commit-Queue: Dan Callaghan <dcallagh@chromium.org>