app-emulation/lxd: Actually install the binaries
Despite the promising name, CROS_GO_BINARIES doesn't do what we need
it to do. Specifically, when we call cros-go_src_compile to build a
package like github.com/lxc/lxd/lxd when in the directory for
github.com/lxc/lxd, it tries to put the lxd binary inside the lxd
sub-directory, and then cros-go_src_install can't find it.

I don't know what the "right" way to do this is, but the easy way is
to go back to what we used to do with 3.17: call "cros_go install" for
each binary package, and then install the binaries from ${GOPATH}/bin,
which avoids any naming collisions.

Because we're now building with go modules, GOPATH can just be a new
empty directory because it's only used to hold these binaries.

BUG=chromium:1295574
TEST=Crostini started with LXD 4.0

Change-Id: I77b8186421ff954a0df23abe18436b256732a53d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3451537
Reviewed-by: Nicholas Verne <nverne@chromium.org>
Commit-Queue: Fergus Dall <sidereal@google.com>
Tested-by: Fergus Dall <sidereal@google.com>
Reviewed-by: David Munro <davidmunro@google.com>