chromeos-base/hps-firmware-tools: build hps-factory
Includes a number of new dependencies, generated by cargo2ebuild.py.

The dev-rust/console ebuild is lightly modified after generation, to fix
a Cargo parse error.

The dev-rust/number_prefix copyright attribution is needed because,
although the license text is in the upstream git repo, it's not included
in the crate.

The dev-rust/simple_logger ebuild includes a trivial upstream patch so
that we can avoid introducing an older copy of dev-rust/colored.

BUG=b:206034105
TEST=FEATURES=test emerge-brya hps-firmware-tools hps-firmware

Change-Id: I9b92382334b8276b88acc7a978fe44e3544e5f5d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3398366
Tested-by: Dan Callaghan <dcallagh@chromium.org>
Reviewed-by: David Lattimore <dml@chromium.org>
Reviewed-by: Sergey Frolov <sfrolov@google.com>
Commit-Queue: Dan Callaghan <dcallagh@chromium.org>