chromeos-chrome: Add hibernate USE flag
Chrome's support of hibernation [1] lives behind a build flag. Add a
corresponding USE flag in the chrome ebuild that a) enables the build
flag in Chrome, and b) adds the hiberman service to the image. This will
be used by the new brya-hibernate builder to produce hibernate-enabled
builds.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/3373198/8

BUG=b:204563517
TEST=None

Change-Id: I54b55ca3cc9df760b7534036d361e224c6b92a90
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3456652
Tested-by: Evan Green <evgreen@chromium.org>
Reviewed-by: Alexis Savery <asavery@chromium.org>
Commit-Queue: Evan Green <evgreen@chromium.org>