chromeos-base/sirenia-tools: don't depend on Rust crates at runtime
This package only installs compiled Rust binaries. The Rust crate
dependencies are only needed for building the binaries, not running
them.

Remove the unneeded RDEPENDS so that we can avoid dragging in a large
number of crate ebuilds when installing the SDK in CI.

BUG=b:219097739
TEST=cros workon --host start sirenia-tools && sudo emerge sirenia-tools

Change-Id: Ia90667bab9cc84c9a52b27bffaf0baba3f092b11
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3457963
Tested-by: Dan Callaghan <dcallagh@chromium.org>
Reviewed-by: Allen Webb <allenwebb@google.com>
Commit-Queue: Dan Callaghan <dcallagh@chromium.org>