virtual/target-chromium-os-dev: skip hps-firmware-tools in asan/ubsan
Rust bare metal targets do not compile successfully in the asan and
ubsan builds, we need to tidy up the per-target rustflags handling
(b/219611035).

Until then, skip building the hps-firmware-tools package. This is not
a big deal because the hps-firmware-tools package isn't included in any
images which are shipped on real devices, only in dev and test images.

BUG=b:206034105
TEST=cros workon --board=brya start virtual/target-chromium-os-dev && \
     emerge-brya virtual/target-chromium-os-dev

Change-Id: If1e50f5309b049e6b700e4065fb2bbba0df7aeb0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3460195
Tested-by: Dan Callaghan <dcallagh@chromium.org>
Reviewed-by: David Lattimore <dml@chromium.org>
Commit-Queue: Dan Callaghan <dcallagh@chromium.org>