cros-go.eclass: cros-go_src_compile install output binary to non-cwd
By default the result binary from the `go build` command is output to
current working directory. When `CROS_WORKON_OUTOFTREE_BUILD=1` is used,
it leaves the binary to the source tree which is not ideal.
Output it to the ${SYSROOT}/tmp/portage/ per-package dir.

BUG=b:220636021
TEST=./build_packages --board=betty --autosetgov

Change-Id: I89d80d16bea8528b4dcfee4c0e4212ddf401874c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3474426
Tested-by: Grace Cham <hscham@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Commit-Queue: Grace Cham <hscham@chromium.org>