shill-cli: remove unused and unfinisehd package
The original author left Google before this effort could yield a
functional dev tool. This package and code should be removed since
nobody is maintaining it and all Connectivity engineers use direct dbus
commands or platform2/shill/test-scripts on a day-to-day basis.

BUG=b:172222930
TEST=The output of $ equery-$BOARD depends shill-cli shows that nothing
depends on this package. It is currently not being build.

Change-Id: I47725ccc8ed9e4ab7fdfdd0826d21504d1378574
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3488246
Tested-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>