cros-kernel2.eclass: Disable FORTIFY_SOURCE for kasan/ubsan
Disable FORTIFY_SOURCE for kasan/ubsan.
Note that this is currently disabled upstream but Chrome OS
has a local revert to have it enabled, see b:207716409.

Since kasan/ubsan are not production configs, it should be
ok to disable it. This fixes the linker error reported
in https://github.com/ClangBuiltLinux/linux/issues/1496
which we also see in kasan/ubsan builds with a newer clang:
https://ci.chromium.org/p/chromeos/builders/cq/amd64-generic-ubsan-cq/b8821193676102987361?

This change can be reverted once there is a general solution
in place for b:207716409.

BUG=b:221611945, b:207716409
TEST=4.14 kernel builds with USE=ubsan and tot llvm.

Change-Id: Ibac6bbf534df9378a4d1315c9b5dde7fe6190871
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3492916
Reviewed-by: Luis Lozano <llozano@chromium.org>
Commit-Queue: Manoj Gupta <manojgupta@chromium.org>
Tested-by: Manoj Gupta <manojgupta@chromium.org>
Auto-Submit: Manoj Gupta <manojgupta@chromium.org>