floss: fix service when /sys/class/bluetooth isn't ready
The btmanagerd service tries to bind-mount /sys/class/bluetooth into the
minijail, which fails if that directory doesn't exist. On the reven
board there are a high number of crash reports for the service failing
to start from this. The /sys/class/bluetooth directory is created lazily
when the driver is ready, so change the bind-mount to /sys/class so that
btmanagerd can access the bluetooth subdirectory when it becomes
available.

BUG=b:221225783
TEST=Run reven on X1 Carbon 6th gen, verify that /sys/class/bluetooth warning no longer shows up in /var/log/bluetooth.log

Change-Id: I52d47ab54475acdb42eaf965e0517f0a82c82cae
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3510298
Tested-by: Nicholas Bishop <nicholasbishop@google.com>
Auto-Submit: Nicholas Bishop <nicholasbishop@google.com>
Reviewed-by: Abhishek Pandit-Subedi <abhishekpandit@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>