net-fs/samba: Uprev samba from 4.15.1 to 4.15.5-r1
Samba versions <4.15.5 have recently been affected by a few CVE's. Uprev
to 4.15.5 requires no dependent package uprevs as they still fall within
the allowed version range.

This removes all the Python dependencies in COMMON_DEPEND as they are
only used for bindings to interface with the Samba server. Given
ChromeOS doesn't use the server portion, it is safe to remove these from
the build file. This was previously a source of divergence from upstream
due to having to rewrite the ebuild.

This version also seems to add AES acceleration [1] through the
cpu_flags_x86_aes library. Prior to this it was not enabled so will
investigate in a follow up CL the impact of this flag but for now it has
no change from the previous behaviour.

[1] https://bugs.gentoo.org/821349

BUG=chromium:1304438, chromium:1293357, chromium:1300507
BUG=chromium:1300508, chromium:1302431, b:221448343
TEST=emerge-{hatch,kevin,kevin64} net-fs/samba
TEST=tast run filemanager.SMB*
TEST=tast run kerberos.*

Change-Id: Iec719042bb834365c52bfea81c9f3386cccfd83e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3534781
Tested-by: Ben Reich <benreich@chromium.org>
Reviewed-by: Josh Simmons <simmonsjosh@google.com>
Reviewed-by: Peter Marshall <petermarshall@chromium.org>
Commit-Queue: Ben Reich <benreich@chromium.org>