dev-go/cncf-xds: remove confusing comment
This comment likely got copied and forgotten when the
ebuild was created. This ebuild is not part of the gcp
collection of modules and is independent of their uprevs
so remove it to avoid confusion.

BUG=b:187795288
TEST=Local builds; CQ.

Change-Id: Ibacc7d13a7c4b92a119a8e29441406d4448f1fe3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3540345
Reviewed-by: Luis Lozano <llozano@chromium.org>
Tested-by: Luis Lozano <llozano@chromium.org>
Commit-Queue: Adrian Ratiu <adrian.ratiu@collabora.corp-partner.google.com>
Tested-by: Adrian Ratiu <adrian.ratiu@collabora.corp-partner.google.com>
Auto-Submit: Adrian Ratiu <adrian.ratiu@collabora.corp-partner.google.com>