arc-cros-gralloc: exclude cros_gralloc.so in ARCVM
The original approach was to exclude this package entirely from the
build. This did not work because it didn't install the test in the
package, despite the test still being valid. The missing test caused
failures (b/222617462).

This approach was originally suggested in http://b/222617462#comment5
Originally we tried `! use video_cards_virgl`, but this evaluated to
false for betty-arc-pi and novato boards, which do use that video card
but run ARC containers.

BUG=b:207805827
TEST=tast run <arc_container_dut> arc.{Gralloc,PlayStore}
TEST=tast run <arc_vm_dut> arc.{Gralloc,PlayStore}.vm
TEST=android-sh -c 'find /vendor -name gralloc.cros.so' (ARC VM dut : no output)

Change-Id: Ice375159aadd24b554f1e1410809a06af2c800da
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3540650
Reviewed-by: Satoshi Niwa <niwa@chromium.org>
Reviewed-by: Yiwei Zhang <zzyiwei@chromium.org>
Commit-Queue: Matthew DeVore <matvore@chromium.org>
Tested-by: Matthew DeVore <matvore@chromium.org>