dev-go/mdns: bump 0.0.1 -> 1.0.5
The mdns version used in CrOS was pretty old and it did
not contain a go.mod definition required for modules
support, so we uprev it to the latest released 1.0.5.

There is no reason to bundle the hashicorp fork of net
because the new mdns depends directly on the upstream
net as well as dns.

BUG=b:187795288
TEST=Local builds; CQ.

Change-Id: Ia9c1f5dcb288d1f6b158f5e225f6e387be624ed7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3545229
Reviewed-by: Luis Lozano <llozano@chromium.org>
Tested-by: Luis Lozano <llozano@chromium.org>
Commit-Queue: Adrian Ratiu <adrian.ratiu@collabora.corp-partner.google.com>
Tested-by: Adrian Ratiu <adrian.ratiu@collabora.corp-partner.google.com>
Auto-Submit: Adrian Ratiu <adrian.ratiu@collabora.corp-partner.google.com>