dev-go/staticcheck: add new package
staticcheck together with the go vet command are the
replacements for the frozen and deprecated golint.

It is also required as a module dependency by some
existing CrOS packages to enable modules mode. For
now we add it as a dependency to dev-go/fscrypt to
ensure it gets built / tested.

A few new packages are added to go-tools which are
required to build staticcheck.

BUG=b:187795288
TEST=Local builds; CQ.

Change-Id: Ifecd2cc8aa21027095c019191cd2477aae75c756
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3545231
Reviewed-by: Luis Lozano <llozano@chromium.org>
Tested-by: Adrian Ratiu <adrian.ratiu@collabora.corp-partner.google.com>
Auto-Submit: Adrian Ratiu <adrian.ratiu@collabora.corp-partner.google.com>
Commit-Queue: Adrian Ratiu <adrian.ratiu@collabora.corp-partner.google.com>