media-libs/mesa-*: drop MESON_AUTO_DEPEND=no
MESON_AUTO_DEPEND was used in the gentoo version of meson.eclass to
allow ebuilds that only conditionally used meson to avoid always
depending on it. It's no longer in the gentoo meson.eclass, and is not
in the chromiumos version either. All of the mesa ebuilds do use meson
for the build, so unconditionally depending on meson is correct.

BUG=b:225350167
TEST=`rg MESON_AUTO_DEP chromiumos/src`, verify no remaining uses
TEST=CQ

Change-Id: If7e058cc314c7d53d392e7736d8e9f77adcdbf79
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3550795
Auto-Submit: Nicholas Bishop <nicholasbishop@google.com>
Reviewed-by: Ilja Friedel <ihf@chromium.org>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>
Tested-by: Nicholas Bishop <nicholasbishop@google.com>