rust: add a rust-host package
Rust is in a weird spot right now: it requires cross-compilers to build,
but we also want to use it to build things for the host. This has led to
quite a bit of pain, so the hope is that we can split Rust into
_either_:
- dev-lang/rust-host & dev-lang/rust, where the former only installs
  artifacts necessary for host binary builds, and the latter installs
  all of the remaining artifacts that dev-lang/rust provides today
- dev-lang/rust for host binaries, and cross-*/rust-libs for each target
  we support

It's unclear what benefits the latter brings over the former, and the
former is simpler to write, maintain, and work with (multiple cross-*
packages will have to deal with mutual exclusion over our shared build/
directory; a single package for all cross-compilers has no such issue),
so do that for now.

Since it's convenient, also include the crrev.com/c/3573236 fix.

BUG=b:227370760, b:227507212
TEST=CQ; emerge rust-host rust; cros_sdk --bootstrap

Change-Id: I7a44a5a71761454aa4801a41c680908b0c37fc9f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3561803
Tested-by: George Burgess <gbiv@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>