swap-init: Get rid of the smaller zram exception for arm
zram needs to be sized to be big enough to handle a compression that's
a bit _better_ that the average comprssion ratio than we expect the
user to encounter. Setting it at the exact average pretty much
guarantees that any time the user happens to have stuff in memory that
compresses a little better then they will be wasting the opportunity
to compress more stuff. It also should be safe to make zram a bit
bigger because it's plausible / realistic for a user to fill memory
with data that can't be compressed, so zram has to deal with this case
anyway. Users deal with big MP4 videos all the time and that's the
definition of big and not compressible.

The old numbers we were using for zram didn't use the above
logic. They were chosen in the early days of zram. Specifically:
* zram was a much less "proven" technology at the time and we didn't
  have lots of experience about how it would handle edge cases.
* zram was lacking a few features that make it work better. "same
  page" detection wasn't added until 2017 and we didn't start using
  multi-stream compression until later as well.
* the kernel's way of dealing with low memory was much less
  mature. The kernel didn't have the OOM reaper (2016) and would often
  get wedged killing tasks.
* userspace's way of dealing with low memory was much less
  mature. We'd _often_ hit the OOM killer in the real world. This is
  much more unusual today and we're better at catching things in
  userspace.
* We didn't have MG-LRU, which makes our memory utilization better.

All of the above point to the fact that we should increase the size of
zram. I've also gathered some data in b/227605780 that confirms it, at
least on trogdor with the platform.MemoryPressure test.

I believe that the sweet spot would be to set this at 2x across the
board. However, as a first step let's just remove the "3 / 2" rule
that's in place for arm. NOTE: the reasons behind the old "3 / 2" rule
are actually still in place, at least for now. ARM GPU (mali and
freedreno) drivers still don't support unpinning, though we're working
on getting this done for freedreno. However, based on the above
argument it still should not be a problem to increase this number and
it allows us to get some data showing the improvement and feels
safe. We'll kick off an experiment to try to move everyone to 2x after
this.

NOTE: for those digging, the very early bug about zram enabling is
crbug.com/218338 from 2012/2013. There is very little justification
about the ARM exception in crrev.com/c/46689 / crrev.com/c/188044 nor
much more in b/35523817. There's a tiny comment in
b/35529264#comment11 from marcheu@, though.

BUG=b:227605780
TEST=platform.MemoryPressure

Change-Id: I94d4293c34bbf67a2b924841ac9b02e713991e5e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3563910
Reviewed-by: Brian Geffon <bgeffon@chromium.org>
Tested-by: Douglas Anderson <dianders@chromium.org>
Commit-Queue: Douglas Anderson <dianders@chromium.org>
Feels: Douglas Anderson <dianders@chromium.org>