chromeos-base/swap-init: Stat the userspace_swap.tmp directory
We will need to expose stateful_partition/unecrypted/userspace_swap.tmp
as a bind mount in debugd so that we can create a temp file. We can
just stat that folder instead.

This CL also introduces a change where we wait up to 5 seconds
for the dm target to be visible. Otherwise when handing the
dm-integrity device to dm-crypt it might return an error.

TEST=Tested on a betty vm
BUG=b:217186049

Change-Id: I58635cd3d2c58c41055866a1326d306c4ab5672a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3563303
Reviewed-by: Sarthak Kukreti <sarthakkukreti@chromium.org>
Tested-by: Brian Geffon <bgeffon@chromium.org>
Auto-Submit: Brian Geffon <bgeffon@chromium.org>
Feels: Brian Geffon <bgeffon@chromium.org>