chromeos-chrome: Remove unnecessary -j argument of ninja with goma
With enabling goma, the two -j arguments are added to the ninja
command line: One is from ${makeopts_jobs} and the other is from the
value calculated in this ebuild.

In that case, the later value (calculated in this file) is effective,
and this doesn't cause a problem actually. But it's confusing so
that this CL removes the former one.

BUG=none
TEST=Only one "-j" argument is included in ninja command with goma

Change-Id: I5769a8f90c13782ea390eb2fc35fd9c37f7e0ee7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3565087
Commit-Queue: Yoshiki Iguchi <yoshiki@chromium.org>
Tested-by: Yoshiki Iguchi <yoshiki@chromium.org>
Auto-Submit: Yoshiki Iguchi <yoshiki@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>