libgrammar: update ebuild
We have successfully use rapid to generate tarball and uploaded to gcs.
This change update the tarball name and folder structure to match with
the auto-released tarball.

For technical reason, version 0.1.0 and 0.2.0 generated by rapid didn't
land in chromeos, so we jump to 0.3.0 directly.

BUG=b:226675194,b:216212852
TEST=tested on zork and verified b/226675194 is fixed

Change-Id: I9016db87e6e87fdb82e63eafb9ec80ade747f7d9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3582029
Reviewed-by: Darren Shen <shend@chromium.org>
Reviewed-by: David Vallet <dvallet@google.com>
Commit-Queue: Jing Wang <jiwan@chromium.org>
Tested-by: Jing Wang <jiwan@chromium.org>