net-wireless: Add default_conn_int USE flag
Some boards require using the default MinConnectionInterval and
MaxConnectionInterval LE parameters in main.conf. Add a USE
flag to comment those lines and use default values. Note to
future writers that this approach is not scalable to solve
generic conf override problems. If we need to add more USE flags
to solve similar issues, a more generic override framework needs
to be developed.

BUG=b:219537522
TEST=cros deploy bluez and verify main.conf changes

Change-Id: I5d240aec5e3cc36eb60dda11252579c2d4a0863c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3570833
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Tested-by: Katherine Lai <laikatherine@chromium.org>
Reviewed-by: Abhishek Pandit-Subedi <abhishekpandit@google.com>