cros-kernel2: Add a new config to support VKMS
The compositor team is currently exploring VKMS as a test backend for
ChromeOS and to support croscomp. This should enable us to test
interactions with the kernel without the need for specialized hardware
as part of CQ.

This should be a no-op, since it defaults to disabled and is not enabled
broadly.

BUG=b:228348590
TEST=Build VM images with USE including 'vkms', and the device is
	available.

Change-Id: Id91cf047e108a437a34dd5bb30e33329b63e202f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3572793
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Auto-Submit: Jim Shargo <jshargo@chromium.org>
Tested-by: Jim Shargo <jshargo@chromium.org>
Commit-Queue: Jim Shargo <jshargo@chromium.org>