dev-lang/ghc: backport upstream AR fix
I've landed this AR fix upstream and this backports
it to our current ghc 8.6.5. Details in the patch file.

This is not enough to switch GHC to build with clang, we
still need to fix the GHC bootstrap tarball for which a
separate bug is created: 219056517.

BUG=b:193847062
TEST=Local builds on x86 & arm; CQ.

Change-Id: Ie3edb2a651b88885adb88e6b3fa7356d6f0c877b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3576433
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Commit-Queue: Manoj Gupta <manojgupta@chromium.org>
Tested-by: Manoj Gupta <manojgupta@chromium.org>